refactor: 폴더 구조 변경에 따른 테스트 클래스 import 경로 수정
